+++ airbyte-integrations/connectors/destination-bigquery/src/test-integration/java/io/airbyte/integrations/destination/bigquery/typing_deduping/BigQuerySqlGeneratorIntegrationTest.java
@@ -4,195 +4,84 @@
 
 package io.airbyte.integrations.destination.bigquery.typing_deduping;
 
+import static com.google.cloud.bigquery.LegacySQLTypeName.legacySQLTypeName;
+import static java.util.stream.Collectors.joining;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.fail;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.cloud.bigquery.BigQuery;
+import com.google.cloud.bigquery.Dataset;
+import com.google.cloud.bigquery.DatasetId;
 import com.google.cloud.bigquery.DatasetInfo;
 import com.google.cloud.bigquery.Field;
 import com.google.cloud.bigquery.FieldValue;
 import com.google.cloud.bigquery.FieldValueList;
 import com.google.cloud.bigquery.QueryJobConfiguration;
 import com.google.cloud.bigquery.Schema;
+import com.google.cloud.bigquery.StandardSQLTypeName;
+import com.google.cloud.bigquery.Table;
 import com.google.cloud.bigquery.TableDefinition;
-import com.google.cloud.bigquery.TableId;
 import com.google.cloud.bigquery.TableResult;
 import io.airbyte.commons.json.Jsons;
-import io.airbyte.integrations.base.destination.typing_deduping.AirbyteProtocolType;
-import io.airbyte.integrations.base.destination.typing_deduping.AirbyteType;
-import io.airbyte.integrations.base.destination.typing_deduping.Array;
-import io.airbyte.integrations.base.destination.typing_deduping.ColumnId;
-import io.airbyte.integrations.base.destination.typing_deduping.StreamConfig;
+import io.airbyte.integrations.base.JavaBaseConstants;
+import io.airbyte.integrations.base.destination.typing_deduping.BaseSqlGeneratorIntegrationTest;
 import io.airbyte.integrations.base.destination.typing_deduping.StreamId;
-import io.airbyte.integrations.base.destination.typing_deduping.Struct;
 import io.airbyte.integrations.destination.bigquery.BigQueryDestination;
-import io.airbyte.protocol.models.v0.DestinationSyncMode;
-import io.airbyte.protocol.models.v0.SyncMode;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
 import org.apache.commons.text.StringSubstitutor;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-// TODO write test case for multi-column PK
 @Execution(ExecutionMode.CONCURRENT)
-public class BigQuerySqlGeneratorIntegrationTest {
+public class BigQuerySqlGeneratorIntegrationTest extends BaseSqlGeneratorIntegrationTest<TableDefinition> {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(BigQuerySqlGeneratorIntegrationTest.class);
-  private static final BigQuerySqlGenerator GENERATOR = new BigQuerySqlGenerator("US");
-  public static final ColumnId ID_COLUMN = GENERATOR.buildColumnId("id");
-  public static final List<ColumnId> PRIMARY_KEY = List.of(ID_COLUMN);
-  public static final ColumnId CURSOR = GENERATOR.buildColumnId("updated_at");
-  public static final ColumnId CDC_CURSOR = GENERATOR.buildColumnId("_ab_cdc_lsn");
-  public static final String QUOTE = "`";
-  private static final LinkedHashMap<ColumnId, AirbyteType> COLUMNS;
-  private static final LinkedHashMap<ColumnId, AirbyteType> CDC_COLUMNS;
 
   private static BigQuery bq;
-  private static BigQueryDestinationHandler destinationHandler;
-
-  private String testDataset;
-  private StreamId streamId;
-
-  static {
-    COLUMNS = new LinkedHashMap<>();
-    COLUMNS.put(ID_COLUMN, AirbyteProtocolType.INTEGER);
-    COLUMNS.put(CURSOR, AirbyteProtocolType.TIMESTAMP_WITH_TIMEZONE);
-    COLUMNS.put(GENERATOR.buildColumnId("struct"), new Struct(new LinkedHashMap<>()));
-    COLUMNS.put(GENERATOR.buildColumnId("array"), new Array(AirbyteProtocolType.UNKNOWN));
-    COLUMNS.put(GENERATOR.buildColumnId("string"), AirbyteProtocolType.STRING);
-    COLUMNS.put(GENERATOR.buildColumnId("number"), AirbyteProtocolType.NUMBER);
-    COLUMNS.put(GENERATOR.buildColumnId("integer"), AirbyteProtocolType.INTEGER);
-    COLUMNS.put(GENERATOR.buildColumnId("boolean"), AirbyteProtocolType.BOOLEAN);
-    COLUMNS.put(GENERATOR.buildColumnId("timestamp_with_timezone"), AirbyteProtocolType.TIMESTAMP_WITH_TIMEZONE);
-    COLUMNS.put(GENERATOR.buildColumnId("timestamp_without_timezone"), AirbyteProtocolType.TIMESTAMP_WITHOUT_TIMEZONE);
-    COLUMNS.put(GENERATOR.buildColumnId("time_with_timezone"), AirbyteProtocolType.TIME_WITH_TIMEZONE);
-    COLUMNS.put(GENERATOR.buildColumnId("time_without_timezone"), AirbyteProtocolType.TIME_WITHOUT_TIMEZONE);
-    COLUMNS.put(GENERATOR.buildColumnId("date"), AirbyteProtocolType.DATE);
-    COLUMNS.put(GENERATOR.buildColumnId("unknown"), AirbyteProtocolType.UNKNOWN);
-
-    CDC_COLUMNS = new LinkedHashMap<>(COLUMNS);
-    CDC_COLUMNS.put(GENERATOR.buildColumnId("_ab_cdc_deleted_at"), AirbyteProtocolType.TIMESTAMP_WITH_TIMEZONE);
-  }
 
   @BeforeAll
-  public static void setup() throws Exception {
+  public static void setupBigquery() throws Exception {
     final String rawConfig = Files.readString(Path.of("secrets/credentials-gcs-staging.json"));
     final JsonNode config = Jsons.deserialize(rawConfig);
-
     bq = BigQueryDestination.getBigQuery(config);
-    destinationHandler = new BigQueryDestinationHandler(bq, "US");
   }
 
-  @BeforeEach
-  public void setupDataset() {
-    testDataset = "bq_sql_generator_test_" + UUID.randomUUID().toString().replace("-", "_");
-    // This is not a typical stream ID would look like, but we're just using this to isolate our tests
-    // to a specific dataset.
-    // In practice, the final table would be testDataset.users, and the raw table would be
-    // airbyte.testDataset_users.
-    streamId = new StreamId(testDataset, "users_final", testDataset, "users_raw", testDataset, "users_final");
-    LOGGER.info("Running in dataset {}", testDataset);
+  @Override
+  protected BigQuerySqlGenerator getSqlGenerator() {
+    return new BigQuerySqlGenerator("US");
+  }
 
-    bq.create(DatasetInfo.newBuilder(testDataset)
-        // This unfortunately doesn't delete the actual dataset after 3 days, but at least we can clear out
-        // the tables if the AfterEach is skipped.
+  @Override
+  protected BigQueryDestinationHandler getDestinationHandler() {
+    return new BigQueryDestinationHandler(bq, "US");
+  }
+
+  @Override
+  protected void createNamespace(String namespace) {
+    bq.create(DatasetInfo.newBuilder(namespace)
+        // This unfortunately doesn't delete the actual dataset after 3 days, but at least we'll clear out old tables automatically
         .setDefaultTableLifetime(Duration.ofDays(3).toMillis())
         .build());
   }
 
-  @AfterEach
-  public void teardownDataset() {
-    bq.delete(testDataset, BigQuery.DatasetDeleteOption.deleteContents());
-  }
-
-  @Test
-  public void softReset() throws InterruptedException {
-    createRawTable();
-    createFinalTableCdc();
+  @Override
+  protected void createRawTable(StreamId streamId) throws InterruptedException {
     bq.query(QueryJobConfiguration.newBuilder(
             new StringSubstitutor(Map.of(
-                "dataset", testDataset)).replace(
+                "raw_table_id", streamId.rawTableId(BigQuerySqlGenerator.QUOTE))).replace(
                 """
-                    ALTER TABLE ${dataset}.users_final ADD COLUMN `weird_new_column` INT64;
-
-                    INSERT INTO ${dataset}.users_raw (`_airbyte_data`, `_airbyte_raw_id`, `_airbyte_extracted_at`) VALUES
-                      (JSON'{"id": 1, "updated_at": "2023-01-01T02:00:00Z", "string": "Alice", "struct": {"city": "San Diego", "state": "CA"}, "integer": 84}', '80c99b54-54b4-43bd-b51b-1f67dafa2c52', '2023-01-01T00:00:00Z'),
-                      (JSON'{"id": 2, "updated_at": "2023-01-01T03:00:00Z", "string": "Bob", "integer": "oops"}', 'ad690bfb-c2c2-4172-bd73-a16c86ccbb67', '2023-01-01T00:00:00Z');
-                    """))
-        .build());
-
-    final String sql = GENERATOR.softReset(incrementalDedupStreamConfig());
-    destinationHandler.execute(sql);
-
-    TableDefinition finalTableDefinition = bq.getTable(TableId.of(testDataset, "users_final")).getDefinition();
-    assertTrue(
-        finalTableDefinition.getSchema().getFields().stream().noneMatch(f -> f.getName().equals("weird_new_column")),
-        "weird_new_column was expected to no longer exist after soft reset");
-    final long finalRows = bq.query(QueryJobConfiguration.newBuilder("SELECT * FROM " + streamId.finalTableId(QUOTE, "")).build()).getTotalRows();
-    assertEquals(2, finalRows);
-    final long rawRows = bq.query(QueryJobConfiguration.newBuilder("SELECT * FROM " + streamId.rawTableId(QUOTE)).build()).getTotalRows();
-    assertEquals(2, rawRows);
-    final long rawUntypedRows = bq.query(QueryJobConfiguration.newBuilder(
-        "SELECT * FROM " + streamId.rawTableId(QUOTE) + " WHERE _airbyte_loaded_at IS NULL").build()).getTotalRows();
-    assertEquals(0, rawUntypedRows);
-  }
-
-  private StreamConfig incrementalDedupStreamConfig() {
-    return new StreamConfig(
-        streamId,
-        SyncMode.INCREMENTAL,
-        DestinationSyncMode.APPEND_DEDUP,
-        PRIMARY_KEY,
-        Optional.of(CURSOR),
-        COLUMNS);
-  }
-
-  private StreamConfig cdcStreamConfig() {
-    return new StreamConfig(
-        streamId,
-        SyncMode.INCREMENTAL,
-        DestinationSyncMode.APPEND_DEDUP,
-        PRIMARY_KEY,
-        // Much like the rest of this class - this is purely for test purposes. Real CDC cursors may not be
-        // exactly the same as this.
-        Optional.of(CDC_CURSOR),
-        CDC_COLUMNS);
-  }
-
-  private StreamConfig cdcIncrementalAppendStreamConfig() {
-    return new StreamConfig(
-        streamId,
-        SyncMode.INCREMENTAL,
-        // This is the only difference between this and cdcStreamConfig.
-        DestinationSyncMode.APPEND,
-        PRIMARY_KEY,
-        Optional.of(CDC_CURSOR),
-        CDC_COLUMNS);
-  }
-
-  // These are known-good methods for doing stuff with bigquery.
-  // Some of them are identical to what the sql generator does, and that's intentional.
-  private void createRawTable() throws InterruptedException {
-    bq.query(QueryJobConfiguration.newBuilder(
-            new StringSubstitutor(Map.of(
-                "dataset", testDataset)).replace(
-                """
-                    CREATE TABLE ${dataset}.users_raw (
+                    CREATE TABLE ${raw_table_id} (
                       _airbyte_raw_id STRING NOT NULL,
                       _airbyte_data JSON NOT NULL,
                       _airbyte_extracted_at TIMESTAMP NOT NULL,
@@ -204,18 +93,22 @@ public class BigQuerySqlGeneratorIntegrationTest {
         .build());
   }
 
-  private void createFinalTableCdc() throws InterruptedException {
+  @Override
+  protected void createFinalTable(boolean includeCdcDeletedAt, StreamId streamId, String suffix) throws InterruptedException {
+    String cdcDeletedAt = includeCdcDeletedAt ? "`_ab_cdc_deleted_at` TIMESTAMP," : "";
     bq.query(QueryJobConfiguration.newBuilder(
             new StringSubstitutor(Map.of(
-                "dataset", testDataset)).replace(
+                "final_table_id", streamId.finalTableId(BigQuerySqlGenerator.QUOTE, suffix),
+                "cdc_deleted_at", cdcDeletedAt)).replace(
                 """
-                    CREATE TABLE ${dataset}.users_final (
+                    CREATE TABLE ${final_table_id} (
                       _airbyte_raw_id STRING NOT NULL,
                       _airbyte_extracted_at TIMESTAMP NOT NULL,
                       _airbyte_meta JSON NOT NULL,
-                      `id` INT64,
-                      `_ab_cdc_deleted_at` TIMESTAMP,
-                      `_ab_cdc_lsn` INT64,
+                      `id1` INT64,
+                      `id2` INT64,
+                      `updated_at` TIMESTAMP,
+                      ${cdc_deleted_at}
                       `struct` JSON,
                       `array` JSON,
                       `string` STRING,
@@ -230,9 +123,232 @@ public class BigQuerySqlGeneratorIntegrationTest {
                       `unknown` JSON
                     )
                     PARTITION BY (DATE_TRUNC(_airbyte_extracted_at, DAY))
-                    CLUSTER BY id, _airbyte_extracted_at;
+                    CLUSTER BY id1, id2, _airbyte_extracted_at;
                     """))
         .build());
+  }
+
+  @Override
+  protected void insertFinalTableRecords(boolean includeCdcDeletedAt, StreamId streamId, String suffix, List<JsonNode> records) throws InterruptedException {
+    List<String> columnNames = includeCdcDeletedAt ? FINAL_TABLE_COLUMN_NAMES_CDC : FINAL_TABLE_COLUMN_NAMES;
+    String cdcDeletedAtDecl = includeCdcDeletedAt ? ",`_ab_cdc_deleted_at` TIMESTAMP" : "";
+    String cdcDeletedAtName = includeCdcDeletedAt ? ",`_ab_cdc_deleted_at`" : "";
+    String recordsText = records.stream()
+        // For each record, convert it to a string like "(rawId, extractedAt, loadedAt, data)"
+        .map(record -> columnNames.stream()
+            .map(record::get)
+            .map(r -> {
+              if (r == null) {
+                return "NULL";
+              }
+              String stringContents;
+              if (r.isTextual()) {
+                stringContents = r.asText();
+              } else {
+                stringContents = r.toString();
+              }
+              return '"' + stringContents
+                  // Serialized json might contain backslashes and double quotes. Escape them.
+                  .replace("\\", "\\\\")
+                  .replace("\"", "\\\"") + '"';
+            })
+            .collect(joining(",")))
+        .map(row -> "(" + row + ")")
+        .collect(joining(","));
+
+    bq.query(QueryJobConfiguration.newBuilder(
+            new StringSubstitutor(Map.of(
+                "final_table_id", streamId.finalTableId(BigQuerySqlGenerator.QUOTE, suffix),
+                "cdc_deleted_at_name", cdcDeletedAtName,
+                "cdc_deleted_at_decl", cdcDeletedAtDecl,
+                "records", recordsText)).replace(
+                // Similar to insertRawTableRecords, some of these columns are declared as string and wrapped in parse_json().
+                // There's also a bunch of casting, because bigquery doesn't coerce strings to e.g. int
+                """
+                    insert into ${final_table_id} (
+                      _airbyte_raw_id,
+                      _airbyte_extracted_at,
+                      _airbyte_meta,
+                      `id1`,
+                      `id2`,
+                      `updated_at`,
+                      `struct`,
+                      `array`,
+                      `string`,
+                      `number`,
+                      `integer`,
+                      `boolean`,
+                      `timestamp_with_timezone`,
+                      `timestamp_without_timezone`,
+                      `time_with_timezone`,
+                      `time_without_timezone`,
+                      `date`,
+                      `unknown`
+                      ${cdc_deleted_at_name}
+                    )
+                    select
+                      _airbyte_raw_id,
+                      _airbyte_extracted_at,
+                      parse_json(_airbyte_meta),
+                      cast(`id1` as int64),
+                      cast(`id2` as int64),
+                      `updated_at`,
+                      parse_json(`struct`),
+                      parse_json(`array`),
+                      `string`,
+                      cast(`number` as numeric),
+                      cast(`integer` as int64),
+                      cast(`boolean` as boolean),
+                      `timestamp_with_timezone`,
+                      `timestamp_without_timezone`,
+                      `time_with_timezone`,
+                      `time_without_timezone`,
+                      `date`,
+                      parse_json(`unknown`)
+                      ${cdc_deleted_at_name}
+                    from unnest([
+                      STRUCT<
+                        _airbyte_raw_id STRING,
+                        _airbyte_extracted_at TIMESTAMP,
+                        _airbyte_meta STRING,
+                        `id1` STRING,
+                        `id2` STRING,
+                        `updated_at` TIMESTAMP,
+                        `struct` STRING,
+                        `array` STRING,
+                        `string` STRING,
+                        `number` STRING,
+                        `integer` STRING,
+                        `boolean` STRING,
+                        `timestamp_with_timezone` TIMESTAMP,
+                        `timestamp_without_timezone` DATETIME,
+                        `time_with_timezone` STRING,
+                        `time_without_timezone` TIME,
+                        `date` DATE,
+                        `unknown` STRING
+                        ${cdc_deleted_at_decl}
+                      >
+                      ${records}
+                    ])
+                    """))
+        .build());
+  }
+
+  @Override
+  protected void insertRawTableRecords(StreamId streamId, List<JsonNode> records) throws InterruptedException {
+    String recordsText = records.stream()
+        // For each record, convert it to a string like "(rawId, extractedAt, loadedAt, data)"
+        .map(record -> JavaBaseConstants.V2_COLUMN_NAMES.stream()
+            .map(record::get)
+            .map(r -> {
+              if (r == null) {
+                return "NULL";
+              }
+              String stringContents;
+              if (r.isTextual()) {
+                stringContents = r.asText();
+              } else {
+                stringContents = r.toString();
+              }
+              return '"' + stringContents
+                  // Serialized json might contain backslashes and double quotes. Escape them.
+                  .replace("\\", "\\\\")
+                  .replace("\"", "\\\"") + '"';
+            })
+            .collect(joining(",")))
+        .map(row -> "(" + row + ")")
+        .collect(joining(","));
+
+    bq.query(QueryJobConfiguration.newBuilder(
+            new StringSubstitutor(Map.of(
+                "raw_table_id", streamId.rawTableId(BigQuerySqlGenerator.QUOTE),
+                "records", recordsText)).replace(
+                // Note the parse_json call, and that _airbyte_data is declared as a string.
+                // This is needed because you can't insert a string literal into a JSON column
+                // so we build a struct literal with a string field, and then parse the field when inserting to the table.
+                """
+                    INSERT INTO ${raw_table_id} (_airbyte_raw_id, _airbyte_extracted_at, _airbyte_loaded_at, _airbyte_data)
+                    SELECT _airbyte_raw_id, _airbyte_extracted_at, _airbyte_loaded_at, parse_json(_airbyte_data) FROM UNNEST([
+                      STRUCT<`_airbyte_raw_id` STRING, `_airbyte_extracted_at` TIMESTAMP, `_airbyte_loaded_at` TIMESTAMP, _airbyte_data STRING>
+                      ${records}
+                    ])
+                    """))
+        .build());
+  }
+
+  @Override
+  protected List<JsonNode> dumpRawTableRecords(StreamId streamId) throws Exception {
+    TableResult result = bq.query(QueryJobConfiguration.of("SELECT * FROM " + streamId.rawTableId(BigQuerySqlGenerator.QUOTE)));
+    return BigQuerySqlGeneratorIntegrationTest.toJsonRecords(result);
+  }
+
+  @Override
+  protected List<JsonNode> dumpFinalTableRecords(StreamId streamId, String suffix) throws Exception {
+    TableResult result = bq.query(QueryJobConfiguration.of("SELECT * FROM " + streamId.finalTableId(BigQuerySqlGenerator.QUOTE, suffix)));
+    return BigQuerySqlGeneratorIntegrationTest.toJsonRecords(result);
+  }
+
+  @Override
+  protected void teardownNamespace(String namespace) {
+    bq.delete(namespace, BigQuery.DatasetDeleteOption.deleteContents());
+  }
+
+  @Test
+  public void testCreateTableIncremental() throws Exception {
+    destinationHandler.execute(generator.createTable(incrementalDedupStream, ""));
+
+    final Table table = bq.getTable(namespace, "users_final");
+    // The table should exist
+    assertNotNull(table);
+    final Schema schema = table.getDefinition().getSchema();
+    // And we should know exactly what columns it contains
+    assertEquals(
+        // Would be nice to assert directly against StandardSQLTypeName, but bigquery returns schemas of
+        // LegacySQLTypeName. So we have to translate.
+        Schema.of(
+            Field.newBuilder("_airbyte_raw_id", legacySQLTypeName(StandardSQLTypeName.STRING)).setMode(Field.Mode.REQUIRED).build(),
+            Field.newBuilder("_airbyte_extracted_at", legacySQLTypeName(StandardSQLTypeName.TIMESTAMP)).setMode(Field.Mode.REQUIRED).build(),
+            Field.newBuilder("_airbyte_meta", legacySQLTypeName(StandardSQLTypeName.JSON)).setMode(Field.Mode.REQUIRED).build(),
+            Field.of("id1", legacySQLTypeName(StandardSQLTypeName.INT64)),
+            Field.of("id2", legacySQLTypeName(StandardSQLTypeName.INT64)),
+            Field.of("updated_at", legacySQLTypeName(StandardSQLTypeName.TIMESTAMP)),
+            Field.of("struct", legacySQLTypeName(StandardSQLTypeName.JSON)),
+            Field.of("array", legacySQLTypeName(StandardSQLTypeName.JSON)),
+            Field.of("string", legacySQLTypeName(StandardSQLTypeName.STRING)),
+            Field.of("number", legacySQLTypeName(StandardSQLTypeName.NUMERIC)),
+            Field.of("integer", legacySQLTypeName(StandardSQLTypeName.INT64)),
+            Field.of("boolean", legacySQLTypeName(StandardSQLTypeName.BOOL)),
+            Field.of("timestamp_with_timezone", legacySQLTypeName(StandardSQLTypeName.TIMESTAMP)),
+            Field.of("timestamp_without_timezone", legacySQLTypeName(StandardSQLTypeName.DATETIME)),
+            Field.of("time_with_timezone", legacySQLTypeName(StandardSQLTypeName.STRING)),
+            Field.of("time_without_timezone", legacySQLTypeName(StandardSQLTypeName.TIME)),
+            Field.of("date", legacySQLTypeName(StandardSQLTypeName.DATE)),
+            Field.of("unknown", legacySQLTypeName(StandardSQLTypeName.JSON))),
+        schema);
+    // TODO this should assert partitioning/clustering configs
+  }
+
+  @Test
+  public void testCreateTableInOtherRegion() throws InterruptedException {
+    BigQueryDestinationHandler destinationHandler = new BigQueryDestinationHandler(bq, "asia-east1");
+    // We're creating the dataset in the wrong location in the @BeforeEach block. Explicitly delete it.
+    bq.getDataset(namespace).delete();
+
+    destinationHandler.execute(new BigQuerySqlGenerator("asia-east1").createTable(incrementalDedupStream, ""));
+
+    // Empirically, it sometimes takes Bigquery nearly 30 seconds to propagate the dataset's existence.
+    // Give ourselves 2 minutes just in case.
+    for (int i = 0; i < 120; i++) {
+      final Dataset dataset = bq.getDataset(DatasetId.of(bq.getOptions().getProjectId(), namespace));
+      if (dataset == null) {
+        LOGGER.info("Sleeping and trying again... ({})", i);
+        Thread.sleep(1000);
+      } else {
+        assertEquals("asia-east1", dataset.getLocation());
+        return;
+      }
+    }
+    fail("Dataset does not exist");
   }
 
   /**
